src/api/pyproject.toml
[project]
name = "open-voice-api"
version = "0.1.0"
description = "Drop-in replacement for OpenAI Realtime API using pluggable ASR/LLM/TTS providers"
requires-python = ">=3.11"
dependencies = [
    "websockets>=13.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0,<1.14.0",
    "python-dotenv>=1.0.0",
    "grpcio>=1.78.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
all-llm = ["anthropic>=0.40.0", "openai>=1.0.0"]
kokoro = ["kokoro-onnx>=0.4.0"]
whisper = ["faster-whisper>=1.0.0"]
edge = ["edge-tts>=6.1.0"]
web-search = ["duckduckgo-search>=7.0.0"]
vad = ["silero-vad>=5.1", "torch>=2.0.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "websockets>=13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
